Extracted the traversers into their own files.
