+++ src/fitnesse/responders/refactoring/SearchReplaceResponder.java
@@ -46,20 +46,3 @@ public class SearchReplaceResponder extends ResultResponder {
 }
 
 
-class SearchReplaceTraverser  implements TraversalListener<WikiPage> {
-
-  private TraversalListener<? super WikiPage> contentReplaceObserver;
-  private TraversalListener<? super WikiPage> webOutputObserver;
-
-  public SearchReplaceTraverser(TraversalListener<? super WikiPage> contentReplaceObserver, TraversalListener<? super WikiPage> webOutputObserver) {
-    this.contentReplaceObserver = contentReplaceObserver;
-    this.webOutputObserver = webOutputObserver;
-  }
-
-  @Override
-  public void process(WikiPage page) {
-    contentReplaceObserver.process(page);
-    webOutputObserver.process(page);
-  }
-
-}
