+++ src/fitnesse/wiki/search/WikiPageFinder.java
@@ -20,20 +20,3 @@ public abstract class WikiPageFinder implements PageFinder {
 }
 
 
-class WikiPageTraverser implements TraversalListener<WikiPage> {
-
-  private final WikiPageFinder finder;
-  private final TraversalListener<? super WikiPage> observer;
-
-  WikiPageTraverser(WikiPageFinder finder, TraversalListener<? super WikiPage> observer){
-    this.finder = finder;
-    this.observer = observer;
-  }
-
-  @Override
-  public void process(WikiPage page) {
-    if (finder.pageMatches(page)) {
-      observer.process(page);
-    }
-  }
-}
